user_auth/requirements.txt
appdirs==1.4.4
argcomplete==1.12.2
asgiref==3.2.3
boto3==1.16.43
botocore==1.19.43
certifi==2020.12.5
cffi==1.14.0
cfgv==3.2.0
cfn-flip==1.2.3
chardet==4.0.0
click==7.1.2
cryptography==2.9.2
distlib==0.3.1
Django==3.0.2
django-cors-headers==3.4.0
django-mysql==3.10.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
durationpy==0.5
filelock==3.0.12
future==0.18.2
hjson==3.0.2
identify==1.5.11
idna==2.10
jmespath==0.10.0
kappa==0.6.0
nodeenv==1.5.0
numpy==1.19.0
pandas==1.0.5
Pillow==7.2.0
pip-tools==5.4.0
pkg-resources==0.0.0
placebo==0.9.0
pre-commit==2.9.3
psycopg2-binary==2.8.6
pushjack==1.6.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.6.1
python-http-client==3.3.1
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.25.1
s3transfer==0.3.3
sendgrid==6.4.4
six==1.15.0
sqlparse==0.3.0
starkbank-ecdsa==1.1.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.55.0
troposphere==2.6.3
twilio==6.46.0
urllib3==1.26.2
virtualenv==20.2.2
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.51.0
